+++ app/src/main/java/com/example/notesandreminding/App.java
@@ -3,19 +3,19 @@ package com.example.notesandreminding;
 import android.app.Application;
 
 public class App extends Application {
-    private static NotesRepository noteRepository;
-    private static SettingsManager settingsManager;
+    private static NotesRepositoryImp noteRepository;
+    private static SettingsManagerImp settingsManager;
     @Override
     public void onCreate() {
         super.onCreate();
 
-        noteRepository = new NotesRepository(this);
-        settingsManager = new SettingsManager(this);
+        noteRepository = new NotesRepositoryImp(this);
+        settingsManager = new SettingsManagerImp(this);
     }
-    public static NotesRepository getNoteRepository() {
+    public static NotesRepositoryImp getNoteRepository() {
         return noteRepository;
     }
-    public static SettingsManager getSettingsManager() {
+    public static SettingsManagerImp getSettingsManager() {
         return settingsManager;
     }
 }
