+++ app/src/main/java/com/example/notesandreminding/NotesRepository.java
@@ -7,89 +7,25 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.concurrent.atomic.AtomicLong;
 
-public class NotesRepository {
-
+public interface NotesRepository {
     ArrayList<Note> notes = new ArrayList<>();
     AtomicLong idCounter = new AtomicLong();
     FileDateSource fds = null;
 
-    public NotesRepository(Context context) {
-        fds = new FileDateSource(context);
+    void sort();
 
-        setNotes(fds.load());
-    }
+    public ArrayList<Note> getNotes();
 
-    private void sort() {
-        Comparator<Note> comNote = new Comparator<Note>() {
-            @Override
-            public int compare(Note left, Note right) {
+    public void removeNote(Long id);
 
-                if (left.getDeadline() == null) {
-                    if (right.getDeadline() != null) {
-                        return 1;
-                    } else {
-                        return Long.compare(right.getEditDate(), left.getEditDate());
-                    }
-                } else {
-                    if (right.getDeadline() == null) {
-                        return -1;
-                    } else {
-                        if (left.getDeadline() < right.getDeadline()) {
-                            return -1;
-                        } else if (left.getDeadline() > right.getDeadline()) {
-                            return 1;
-                        } else {
-                            return Long.compare(right.getEditDate(), left.getEditDate());
-                        }
-                    }
-                }
-            }
-        };
+    public void setNotes(ArrayList<Note> notes);
 
-        Collections.sort(notes, comNote);
-    }
+    public void saveNote(Note n);
 
-    public ArrayList<Note> getNotes() {
-        return notes;
-    }
+    public Note getNote(Long id);
 
-    public void removeNote(Long id) {
-        notes.remove(getNote(id));
-    }
+    public Long createID();
 
-    public void setNotes(ArrayList<Note> notes) {
-        this.notes = notes;
-    }
-
-    public void saveNote(Note n) {
-        if (getNote(n.getId()) != null) {
-            removeNote(n.getId());
-            notes.add(n);
-        } else {
-            n.setId(createID());
-            notes.add(n);
-        }
-        sort();
-        saveToDisk();
-    }
-
-
-    public Note getNote(Long id) {
-        Note n = null;
-        for (int i = 0; i < notes.size(); i++) {
-            if (notes.get(i).getId() == id) {
-                n = notes.get(i);
-            }
-        }
-        return n;
-    }
-
-    public synchronized Long createID() {
-        return idCounter.getAndIncrement();
-    }
-
-    private void saveToDisk() {
-        boolean result = fds.save(notes);
-    }
+    void saveToDisk();
 
 }
